StreamLit/requirements.txt
streamlit
pandas
sqlalchemy
google-generativeai
paramiko
sshtunnel
beautifulsoup4
lxml
requests
pdfminer.six
streamlit-authenticator
pyyaml
pysqlcipher3
cryptography
requests